QSP-393 Codestyle proposal and static analysis bugs (#24)
* Organized imports

* Reformatted the code

* Removed audit module bugs. Includes:

- Removing a constructor bug in AnalyzerRunException
- Renaming exec methods that override built-ins
- Fixing some typos

* Polished config code. Includes:

- Breaking too long docstrings
- Removing unused imports
- Replacing != None with with is not None
- Removing a semicolon
- Removing redundant parentheses

* Debugged evt module. Includes:

- Removing shadow built-ins
- Fixing unbound variables in except blocks

* Removed unused import from streaming module.

* Removed semicolon from upload module.

* Debugged utils.db module. Includes:

- Replacing comparison to None using is instead ==
- Removing tautology in sql3lite

* Polished the utils core. Includes:

- Removing parentheses
- Removing unused imports

* Polished the main class. Includes:

- Removing shadows of outside of scope signal
- Removing unused reference
- Fixing unbound variables

* Reformatted the util.eth module

* Documented brief codestyle instructions.

* Removed obsolete todo.

* Removed indentation bug

* Incorporated the code review. Includes:

- Removing a typo bug
- Removing finally block
- Printing into stderr if logger is not available

* Adjusted indentation and renamed a variable

* Addressed more of static analysis and codestyle:

- Removed mutable default parameter
- Fixed indentation and whitespace
- Addressed unbound variables
- Removed unused imports
- Equipped generic except

+++ qsp_protocol_node/evt/evt_pool_manager.py
@@ -1,37 +1,38 @@
-import apsw
 import os
 
 from pathlib import Path
-from utils.db import Sqlite3Worker, get_first
+from utils.db import Sqlite3Worker
+from utils.db import get_first
 
-class EventPoolManager:    
+
+class EventPoolManager:
     @staticmethod
-    def __encode(dict):
-        if dict is None:
+    def __encode(dictionary):
+        if dictionary is None:
             return None
 
-        new_dict = {}
-        for key in dict.keys():
+        new_dictionary = {}
+        for key in dictionary.keys():
             if key == "price" or key == "block_nbr":
-                new_dict[key] = str(dict[key])
+                new_dictionary[key] = str(dictionary[key])
             else:
-                new_dict[key] = dict[key]
+                new_dictionary[key] = dictionary[key]
 
-        return new_dict
+        return new_dictionary
 
     @staticmethod
-    def __decode(dict):
-        if dict is None:
+    def __decode(dictionary):
+        if dictionary is None:
             return None
 
-        new_dict = {}
-        for key in dict.keys():
+        new_dictionary = {}
+        for key in dictionary.keys():
             if key == "price" or key == "block_nbr":
-                new_dict[key] = int(dict[key])
+                new_dictionary[key] = int(dictionary[key])
             else:
-                new_dict[key] = dict[key]
+                new_dictionary[key] = dictionary[key]
 
-        return new_dict
+        return new_dictionary
 
     @staticmethod
     def __query_path(query):
@@ -56,6 +57,7 @@ class EventPoolManager:
         error = False
 
         self.__sqlworker = None
+        db_file = None
         try:
             db_file = Path(db_path)
             if db_file.is_file():
@@ -66,17 +68,17 @@ class EventPoolManager:
 
             if not db_existed:
                 EventPoolManager.__exec_sql(self.__sqlworker, 'createdb')
-        
+
         except Exception:
             error = True
             raise
-        
+
         finally:
             if error:
                 if self.__sqlworker is not None:
                     self.__sqlworker.close()
 
-                if not db_existed and db_created:
+                if not db_existed and db_created and db_file is not None:
                     db_file.unlink()
 
     @property
@@ -89,7 +91,7 @@ class EventPoolManager:
 
     def is_request_processed(self, request_id):
         row = self.get_event_by_request_id(request_id)
-        return not(row is None or row == {})
+        return not (row is None or row == {})
 
     def get_next_block_number(self):
         current = self.get_latest_block_number()
@@ -117,13 +119,18 @@ class EventPoolManager:
             )
         )
 
-    def __process_evt_with_status(self, query_name, fct, values=(), fct_kwargs={}):
+    def __process_evt_with_status(self, query_name, fct, values=(), fct_kwargs=None):
         for evt in EventPoolManager.__exec_sql(self.__sqlworker, query_name, values):
             decoded_evt = EventPoolManager.__decode(evt)
-            fct(decoded_evt, **fct_kwargs)
+            if fct_kwargs is None:
+                fct(decoded_evt, {})
+            else:
+                fct(decoded_evt, **fct_kwargs)
 
     def get_event_by_request_id(self, request_id):
-        row = get_first(EventPoolManager.__exec_sql(self.__sqlworker, 'get_event_by_request_id', (request_id,)))
+        rows = EventPoolManager.__exec_sql(self.__sqlworker, 'get_event_by_request_id',
+                                           (request_id,))
+        row = get_first(rows)
         return EventPoolManager.__decode(row)
 
     def process_incoming_events(self, process_fct):
@@ -150,7 +157,7 @@ class EventPoolManager:
         encoded_evt = EventPoolManager.__encode(evt)
         EventPoolManager.__exec_sql(
             self.__sqlworker,
-           'set_evt_to_assigned',
+            'set_evt_to_assigned',
             (encoded_evt['evt_name'], encoded_evt['status_info'], encoded_evt['request_id'],),
         )
 
@@ -160,12 +167,12 @@ class EventPoolManager:
             self.__sqlworker,
             'set_evt_to_be_submitted',
             (encoded_evt['status_info'],
-                encoded_evt['tx_hash'],
-                encoded_evt['report_uri'],
-                encoded_evt['report_hash'],
-                encoded_evt['audit_state'],
-                encoded_evt['request_id'],
-            ),
+             encoded_evt['tx_hash'],
+             encoded_evt['report_uri'],
+             encoded_evt['report_hash'],
+             encoded_evt['audit_state'],
+             encoded_evt['request_id'],
+             ),
         )
 
     def set_evt_to_submitted(self, evt):
@@ -174,19 +181,19 @@ class EventPoolManager:
             self.__sqlworker,
             'set_evt_to_submitted',
             (encoded_evt['tx_hash'],
-                encoded_evt['status_info'],
-                encoded_evt['report_uri'],
-                encoded_evt['report_hash'],
-                encoded_evt['audit_state'],
-                encoded_evt['request_id'],
-            ),
+             encoded_evt['status_info'],
+             encoded_evt['report_uri'],
+             encoded_evt['report_hash'],
+             encoded_evt['audit_state'],
+             encoded_evt['request_id'],
+             ),
         )
 
     def set_evt_to_done(self, evt):
         encoded_evt = EventPoolManager.__encode(evt)
         EventPoolManager.__exec_sql(
             self.__sqlworker,
-           'set_evt_to_done',
+            'set_evt_to_done',
             (encoded_evt['status_info'], encoded_evt['request_id'],),
         )
